refactor: migrate to genai-lite LLM module
Replace integrated LLM implementation with genai-lite package:
- Add genai-lite ^0.1.0 dependency
- Remove old electron/modules/llm directory
- Implement custom ApiKeyProvider using genai-key-storage-lite
- Centralize IPC channel constants in common/types/llm.ts
- Update all imports and type references to use genai-lite
- Fix webpack configs to include common directory
- Maintain full backward compatibility with existing API
