+++ package.json
@@ -107,6 +107,7 @@
     "diff-match-patch": "^1.0.5",
     "fix-path": "^4.0.0",
     "genai-key-storage-lite": "^0.1.4",
+    "genai-lite": "^0.1.0",
     "ignore": "^7.0.0",
     "js-tiktoken": "^1.0.16",
     "lucide-react": "^0.469.0",
